batch processing with glue
